test: unify e2e utils across e2e/deployment/openshift tests
